+++ functions/Get-PSBlueskyBlockList.ps1
@@ -54,6 +54,7 @@ Function Get-BskyBlockedList {
 
             $results = @()
             $response = Invoke-RestMethod -Uri $apiUrl -Method Get -Headers $headers -ResponseHeadersVariable rh
+            _newLogData -apiUrl $apiUrl -command $MyInvocation.MyCommand | _updateLog
             Write-Information -MessageData $rh -Tags ResponseHeader
             If ($response) {
                 $results += $response.lists
@@ -64,6 +65,7 @@ Function Get-BskyBlockedList {
                     while ($response.cursor) {
                         $url = $apiUrl + "&cursor=$($response.cursor)"
                         $response = Invoke-RestMethod -Uri $url -Method Get -Headers $headers
+                        _newLogData -apiUrl $apiUrl -command $MyInvocation.MyCommand | _updateLog
                         If ($response.followers) {
                             Write-Information -MessageData $response -Tags raw
                             $results += $response.followers