+++ functions/psblueskychat.ps1
@@ -1,3 +1,4 @@
+# IN DEVELOPMENT
 
 #https://docs.bsky.app/docs/api/chat-bsky-convo-list-convos
 #https://docs.bsky.app/docs/api/chat-bsky-convo-get-convo